testejä ja dokumentaatiota

+++ src/services/edit_service.py
@@ -2,7 +2,6 @@ import copy
 from entities.puzzle import Puzzle
 from entities.quiz import Quiz
 from repositories.quiz_repository import QuizRepository
-from services.management_service import management_service
 
 
 class EditService:
@@ -12,7 +11,7 @@ class EditService:
         tilasta eli esimerkiksi siitä, mitä arvoituksia kyseiseen visailuun on liitetty.
     """
 
-    def __init__(self, old_quiz: Quiz):
+    def __init__(self, old_quiz: Quiz, repository: QuizRepository):
         """Luokan konstruktori. Luo uuden visailujen muokkaamislogiikasta vastaavan palvelun.
 
         Args:
@@ -20,7 +19,7 @@ class EditService:
                                 ensin kopion, johon muutokset tehdään siltä varalta, että 
                                 käyttäjä haluaakin peruuttaa tehdyt muutokset.
         """
-        self._repository = QuizRepository()
+        self._repository = repository
         self._old_quiz = old_quiz
         self._quiz = copy.deepcopy(old_quiz)
 
@@ -47,11 +46,11 @@ class EditService:
         """
         name = name.strip()
 
-        if name == "":
+        if not name:
             raise NoQuizNameError("Quiz name is missing.")
 
         old_name = self._old_quiz.name
-        quiz_names = management_service.get_quiz_names()
+        quiz_names = self._repository.find_quiz_names()
 
         if name in quiz_names and name != old_name:
             raise QuizExistError("A quiz with same name already exists.")
@@ -66,17 +65,31 @@ class EditService:
         """
         return self._quiz.puzzles
 
-    def add_puzzle(self, user_input):
+    def add_puzzle(self, name, words):
         """Lisää luotavaan visailuun uuden arvoituksen käyttäjän syötteen perusteella.
 
         Args:
             user_input (list): Lista, joka sisältää käyttäjän PuzzleCreationWindow-
                                näkymässä antaman syötteen.
         """
-        name = user_input.pop(0)
-        words = user_input
-        puzzle = Puzzle(name, words)
+        name_char_limit = 100
+        word_char_limit = 12
 
+        if not name.strip():
+            raise NoPuzzleNameError("Puzzle name is missing.")
+        if len(name.strip()) > name_char_limit:
+            raise PuzzleNameLenghtError(f"Puzzle name is too long (limit {name_char_limit} characters)")
+
+        for word in words:
+            if not word.strip():
+                raise NoPuzzleWordError("One ore more puzzle words is missing.")
+            if " " in word.strip():
+                raise MultipleWordsError("Only one word per a word field is allowed.")
+            char_limit = 12
+            if len(word.strip()) > word_char_limit:
+                raise PuzzleWordLenghtError(f"One ore more puzzle words are too long (limit {char_limit} characters)")
+            
+        puzzle = Puzzle(name, words)
         self._quiz.puzzles.append(puzzle)
 
     def remove_puzzle(self, index):
@@ -89,7 +102,10 @@ class EditService:
 
     def save_quiz(self):
         """Tallentaa luotavana olevan visailun tietokantaan."""
-        
+
+        if not self._quiz.puzzles:
+            raise NoPuzzlesError("Not allowed to save a quiz with zero puzzles.")
+
         if self._old_quiz.name != "":
             self._repository.delete_quiz(self._old_quiz)
 
@@ -100,4 +116,22 @@ class QuizExistError(Exception):
     pass
 
 class NoQuizNameError(Exception):
+    pass
+
+class NoPuzzlesError(Exception):
+    pass
+
+class NoPuzzleNameError(Exception):
+    pass
+
+class NoPuzzleWordError(Exception):
+    pass
+
+class MultipleWordsError(Exception):
+    pass
+
+class PuzzleNameLenghtError(Exception):
+    pass
+
+class PuzzleWordLenghtError(Exception):
     pass
